更改禁止调用alloc等方法初始化

+++ test/分层日志/封装/XYSLogFormat.m
@@ -44,7 +44,7 @@
     //取得行数
     NSUInteger numLine = (unsigned long)logMessage->_line;
     
-    NSString *formatString = [NSString stringWithFormat:@"%@ %@[%@] %@ %@ %lu\n\n",currentDateString,content,thredId,funcTionName,methodName,numLine];
+    NSString *formatString = [NSString stringWithFormat:@"%@ %@[%@] %@ %@ %lu\n\n",currentDateString,content,thredId,funcTionName,methodName,(unsigned long)numLine];
     
     return formatString;
